require authentication to make changes to a group

+++ server/schemas/resolvers.js
@@ -69,18 +69,38 @@ const resolvers = {
       return { token, user };
     },
     createGroup: async (parent, { name, admin }, context) => {
-      const groupData = await Group.create({ name: name, admins: [admin] });
+      const groupData = await Group.create({
+        name: name,
+        admins: [admin.user.id],
+      });
       return groupData;
     },
-    updateGroup: async (parent, { groupId, name }, context) => {
-      const groupData = await Group.findByIdAndUpdate(groupId, { name: name });
-      return groupData;
-    },
-    deleteGroup: async (parent, { groupId }, context) => {
-      const groupData = await Group.findByIdAndDelete(groupId);
-    },
-    addGroupMember: async (parent, { userId, groupId }, context) => {
+    updateGroup: async (parent, { groupId, name, admin }, context) => {
       const groupData = await Group.findById(groupId);
+      let isAdmin = groupData.admins.find(
+        (element) => element === admin.user.id
+      );
+      if (!isAdmin) return;
+      groupData.name = name;
+      groupData.save();
+      return groupData;
+    },
+    // TODO: Delete group from all users' groups array
+    deleteGroup: async (parent, { groupId, admin }, context) => {
+      const groupData = await Group.findById(groupId);
+      let isAdmin = groupData.admins.find(
+        (element) => element === admin.user.id
+      );
+      if (!isAdmin) return;
+      const groupDelete = await Group.findByIdAndDelete(groupId);
+      return groupDelete;
+    },
+    addGroupMember: async (parent, { userId, groupId, admin }, context) => {
+      const groupData = await Group.findById(groupId);
+      let isAdmin = groupData.admins.find(
+        (element) => element === admin.user.id
+      );
+      if (!isAdmin) return;
       if (groupData.members.find((element) => element.user === userId)) {
         return false;
       }
@@ -90,6 +110,10 @@ const resolvers = {
     },
     removeGroupMember: async (parent, { userId, groupId }, context) => {
       const groupData = await Group.findById(groupId);
+      let isAdmin = groupData.admins.find(
+        (element) => element === admin.user.id
+      );
+      if (!isAdmin) return;
       let index = groupData.members.findIndex(
         (element) => element.user === userId
       );
